# specialize io::copy to use copy_file_range, splice or sendfile