@KodrAus 2020-09-04
Thanks for your patience on this so far @the8472! Since this is quite a large PR it'll take me a while to dig through it all.

It's tricky to benchmark IO code, but I would be interested to see roughly what the impact of this is so we can justify the new machinery a bit.