@the8472 2020-09-04
@KodrAus it's not just about speed (but that ought to be a benefit too, sure) but also about using `copy_file_range` which reduces storage footprint on copy-on-write filesystems.

But I'll try to come up with a benchmark.

That said, I think #75428, being a bugfix, has a higher priority than this.