@the8472 2020-09-05
@KodrAus 

```
ext4
[old] test io::tests::bench_file_to_file_copy         ... bench:      82,997 ns/iter (+/- 7,423) = 1579 MB/s
[new] test io::tests::bench_file_to_file_copy         ... bench:      65,165 ns/iter (+/- 2,448) = 2011 MB/s

btrfs
[old] test io::tests::bench_file_to_file_copy         ... bench:      75,250 ns/iter (+/- 2,551) = 1741 MB/s
[new] test io::tests::bench_file_to_file_copy         ... bench:      10,802 ns/iter (+/- 363) = 12134 MB/s


[old] test io::tests::bench_file_to_pipe_copy         ... bench:      37,602 ns/iter (+/- 19,304) = 3485 MB/s
[new] test io::tests::bench_file_to_pipe_copy         ... bench:      31,220 ns/iter (+/- 15,320) = 4198 MB/s

[old] test io::tests::bench_pipe_to_socket_copy       ... bench:      79,906 ns/iter (+/- 5,964) = 1640 MB/s
[new] test io::tests::bench_pipe_to_socket_copy       ... bench:      54,283 ns/iter (+/- 15,867) = 2414 MB/s
```

All benchmarks copy 128KiB per bencher iteration.  Larger files will probably achieve higher speeds.


While benchmarking I have noticed that results involving pipes can be quite erratic since currently the only form of pipes exposed in the public API is for child process stdio, which means the performance characteristics depend on the process on the other end.
I have used `yes` and `dd` to get the loop of syscalls that I want, but even minor changes can mean one process overtakes another which leads to blocking IO calls suddenly showing drastically different performance. The generic `io::copy` loop uses 1 read and 1 write syscall with an 8k buffer each. While that is in principle slower it can occasionally show higher throughput in benchmarks because `splice` finishes so fast that the source/sink becomes empty/full and the process gets suspended and incurs additional context switch costs while the constant switching between read/write means the process never gets suspended because it is the limiting factor.
`sendfile` from file to pipe works on modern kernels and doesn't seem to suffer from this erratic behavior because it stays in kernel space and just waits when the pipe is full, thus avoiding the context switch overhead.

So I'll go ahead and change the behavior to do syscall bashing to figure out which one works on a pair of file descriptors, with this preference ranking: *copy_file_range > sendfile > splice > read+write*
Modern kernels should then mostly succeed with the first two which provide the most benefits while older kernels will hit the fallbacks more frequently.