@the8472 2020-09-10
@KodrAus updated and shaved off some syscalls. You can find benchmarks on top.