@the8472 2020-10-12
@KodrAus 

> I think the module structure could look something like this: [...]

The issue is that there are more implementations of `copy(from: &Path, to: &Path)` in the `unix::fs` module. E.g. a macos one. Should that also be moved into the new `unix::copy` module?