@the8472 2020-10-15
> That way none of the CopyResult or other internal enums need to be visible in the rest of the crate. What do you think?

I have restricted the visibility of those without reshuffling the other copy methods already present in the `unix::fs` module.

But I'm still lacking a better name for than `generic_copy` or `copy_fallback`. 
Perhaps `userspace_copy`? But that's from the perspective of the specializations that can do kernel offload, not from the perspective of generic `Read` and `Write`.