@the8472 2020-10-15
Actually, I can move the linux-only `copy_file_range/sendfile/splice` methods while leaving the multi-platform `copy` wrappers where they are.