@hyd-dev 2020-11-06
Android link failure:
```
error: linking with `/android/ndk/arm-14/bin/arm-linux-androideabi-clang` failed: exit code: 1
```
```
std::sys::unix::kernel_copy::sendfile_splice::h9b10b6580108736a: error: undefined reference to 'splice'
```

@the8472 I think my #74426 can be closed after merging this. If you want, you can add "Closes #74426" to the PR description. Thank you!