@the8472 2020-11-06
I seems like the android build failure is because the symbol is only exposed for android [API level 21 or higher](https://github.com/aosp-mirror/platform_bionic/commit/065237718a557ac47e9a03521a2d52330199cba8) while CI builds against [API level 14](https://github.com/rust-lang/rust/blob/a601302ff0217b91589b5a7310a8a23adb843fdc/src/ci/docker/host-x86_64/arm-android/Dockerfile#L8)

I'll try the raw syscall then.