@the8472 2020-11-07
Added some fixes for android, `./src/ci/docker/run.sh arm-android` passes locally. A side-benefit is that now it won't waste syscalls if seccomp blocks sendfile or splice.