@KodrAus 2020-11-08
@the8472 Do you know if we have precedent elsewhere in the standard library for probing syscall availability like this (outside of Windows, where I think it's pretty standard)? It seems slightly suspect to me. If it's already done in a similar way then I'm ok with it, otherwise maybe we should just remove `android` from the`cfg`s.