@the8472 2020-11-08
@KodrAus sure, here's a macos example: https://github.com/rust-lang/rust/blob/master/library/std/src/sys/unix/fs.rs#L1327-L1337
and there are a few places where it's done on linux too.