@mati865 2020-11-08
Quite possible it's spurious: https://rust-lang.zulipchat.com/#narrow/stream/242791-t-infra/topic/segfaults